% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLRMC.R
\name{ECLRMC}
\alias{ECLRMC}
\title{Takes an incomplete matrix and returns the imputed matrix using ECLRMC method.}
\usage{
ECLRMC(x, beta = 0.1)
}
\arguments{
\item{x}{An m by n matrix with NAs}

\item{beta}{A value in [0,1] range. Higher beta value means comparing each row with more nearest neighbours. Default value = 0.1}
}
\value{
An m by n matrix with imputed values
}
\description{
Takes an incomplete matrix and returns the imputed matrix using ECLRMC method.
}
\examples{
x = matrix(c(5.1, 4.9, NA, 4.6, 3.5, 3.0, 3.2, 3.1, 1.4, NA, 1.3, 1.5), byrow = TRUE, ncol=4)
ECLRMC(x, beta = 0.2)
}
\references{
Chen, Xiaobo, et al. "Ensemble correlation-based low-rank matrix completion with applications to traffic data imputation." Knowledge-Based Systems 132 (2017): 249-262.
}
