% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLRMC.R
\name{LRMC}
\alias{LRMC}
\title{Takes an incomplete matrix and returns the imputed matrix using LRMC method.}
\usage{
LRMC(x)
}
\arguments{
\item{x}{An m by n matrix with NAs}
}
\value{
An m by n matrix with imputed values
}
\description{
Takes an incomplete matrix and returns the imputed matrix using LRMC method.
}
\examples{
x = matrix(c(5.1, 4.9, NA, 4.6, 3.5, 3.0, 3.2, 3.1, 1.4, NA, 1.3, 1.5), byrow = TRUE, ncol=4)
LRMC(x)
}
\references{
Chen, Xiaobo, et al. "Ensemble correlation-based low-rank matrix completion with applications to traffic data imputation." Knowledge-Based Systems 132 (2017): 249-262.
}
